+++ objetivo1_lista1/src/model/Funcionario.java
@@ -2,24 +2,24 @@ package model;
 
 public class Funcionario {
 
-    private int id;
+    private Long id;
     private String nome;
     private double salario;
 
     public Funcionario() {
     }
 
-    public Funcionario(int id, String nome, double salario) {
+    public Funcionario(Long id, String nome, double salario) {
         this.id = id;
         this.nome = nome;
         this.salario = salario;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
